# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BAM Store API",
        "description": "API Documentation for BAM Store",
        "contact": {
            "email": "admin@bamstore.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/admin/products": {
            "get": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Get list of products (Admin)",
                "operationId": "2c536fdba30bf94d8f127e47ab617917",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Create a new product",
                "operationId": "e048ab31940c05ee2bf01680df6f517c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category_id",
                                    "name",
                                    "status",
                                    "base_price"
                                ],
                                "properties": {
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "vendor_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "New Product"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Product description"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ready",
                                            "pre_order"
                                        ],
                                        "example": "ready"
                                    },
                                    "base_price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 10000
                                    },
                                    "image_url": {
                                        "type": "string"
                                    },
                                    "payload": {
                                        "description": "JSON string for complex nested data (variants, skus)",
                                        "type": "string"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/products/{id}": {
            "get": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Get product details (Admin)",
                "operationId": "28af906b711dfb80b17d1f59bb42f9b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Update a product",
                "description": "Use POST with _method=PUT or just POST for multipart/form-data update",
                "operationId": "59e2f7e48a149f2b6250cc0ca12bb570",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "category_id": {
                                        "type": "integer"
                                    },
                                    "vendor_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "ready",
                                            "pre_order"
                                        ]
                                    },
                                    "base_price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "payload": {
                                        "description": "JSON string for complex nested data",
                                        "type": "string"
                                    },
                                    "images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "binary"
                                        }
                                    },
                                    "deleted_images[]": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Products"
                ],
                "summary": "Delete a product",
                "operationId": "808e2a204827a8ad16c0abd889faa731",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID or Slug",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}